+++ R/pluck.R
@@ -41,7 +41,6 @@
 #' obj2 <- list("b", list(2, elt = "bar"))
 #' x <- list(obj1, obj2)
 #'
-#'
 #' # pluck() provides a way of retrieving objects from such data
 #' # structures using a combination of numeric positions, vector or
 #' # list names, and accessor functions.
@@ -75,37 +74,12 @@
 #' try(chuck(x, 10))
 #' try(chuck(x, 1, 10))
 #'
-#'
 #' # The map() functions use pluck() by default to retrieve multiple
 #' # values from a list:
 #' map(x, 2)
 #'
-#' # Pass multiple indexes with a list:
-#' map(x, list(2, "elt"))
-#'
-#' # This is equivalent to:
-#' map(x, pluck, 2, "elt")
-#'
-#' # You can also supply a default:
-#' map(x, list(2, "elt", 10), .default = "superb default")
-#'
-#' # Or use the strict variant:
-#' try(map(x, chuck, 2, "elt", 10))
-#'
-#'
-#' # You can also assign a value in a pluck location with pluck<-:
-#' pluck(x, 2, 2, "elt") <- "quuux"
-#' x
-#'
-#' # This is a shortcut for the prefix function assign_in():
-#' y <- assign_in(x, list(2, 2, "elt"), value = "QUUUX")
-#' y
-#'
-#'
 #' # pluck() also supports accessor functions:
 #' my_element <- function(x) x[[2]]$elt
-#'
-#' # The accessor can then be passed to pluck:
 #' pluck(x, 1, my_element)
 #' pluck(x, 2, my_element)
 #'
@@ -114,7 +88,6 @@
 #' # right-to-left and from left-to-right in different parts of the
 #' # expression:
 #' my_element(x[[1]])
-#'
 #'
 #' # If you have a list of accessors, you can splice those in with `!!!`:
 #' idx <- list(1, my_element)
@@ -210,4 +183,64 @@ is_index <- function(x) {
 attr_getter <- function(attr) {
   force(attr)
   function(x) attr(x, attr, exact = TRUE)
+}
+
+
+#' Modify a pluck location
+#'
+#' @description
+#'
+#' * `assign_in()` takes a data structure and a [pluck] location,
+#'   assigns a value there, and returns the modified data structure.
+#'
+#' * `modify_in()` applies a function to a pluck location, assigns the
+#'   result back to that location with [assign_in()], and returns the
+#'   modified data structure.
+#'
+#' @inheritParams pluck
+#' @param .f A function to apply at the pluck location given by `.where`.
+#' @param ... Arguments passed to `.f`.
+#' @param .where,where A pluck location, as a numeric vector of
+#'   positions, a character vector of names, or a list combining both.
+#'   The location must exist in the data structure.
+#' @seealso [pluck()]
+#' @export
+#' @examples
+#' # Recall that pluck() returns a component of a data structure that
+#' # might be arbitrarily deep
+#' x <- list(list(bar = 1, foo = 2))
+#' pluck(x, 1, "foo")
+#'
+#' # Use assign_in() to modify the pluck location:
+#' str(assign_in(x, list(1, "foo"), 100))
+#'
+#' # Like pluck(), this works even when the element (or its parents) don't exist
+#' pluck(x, 1, "baz")
+#' str(assign_in(x, list(2, "baz"), 100))
+#'
+#' # modify_in() applies a function to that location and update the
+#' # element in place:
+#' modify_in(x, list(1, "foo"), ~ .x * 200)
+#'
+#' # Additional arguments are passed to the function in the ordinary way:
+#' modify_in(x, list(1, "foo"), `+`, 100)
+modify_in <- function(.x, .where, .f, ...) {
+  .where <- as.list(.where)
+  .f <- rlang::as_function(.f)
+
+  value <- .f(pluck(.x, !!!.where), ...)
+  assign_in(.x, .where, value)
+}
+#' @rdname modify_in
+#' @param value A value to replace in `.x` at the pluck location.
+#' @export
+assign_in <- function(x, where, value) {
+  n <- length(where)
+  if (n == 0) {
+    abort("`where` must contain at least one element")
+  } else if (n > 1) {
+    value <- assign_in(pluck(x, where[[1]]), where[-1], value)
+  }
+  list_slice2(x, where[[1]]) <- value
+  x
 }